I hope this catches the compile errors

+++ include/wifi_drv.h
@@ -41,7 +41,7 @@ void WiFiDrv_wifiDriverInit(void);
 
 void WiFiDrv_wifiDriverDeinit(void);
 
-int WiFiDrv_getRemoteData(uint8 sock, uint8 *ip, uint8 *port);
+int WiFiDrv_getRemoteData(uint8 sock, uint32 *ip, uint32 *port);
 
 /*
  * Set the desired network which the connection manager should try to
